[Refactor] 강연/세미나 기록장 기록하기 리팩토링
