--- src/main/java/careerfestival/career/domain/Comment.java
@@ -1,30 +1,0 @@
-package careerfestival.career.domain;
-
-import careerfestival.career.domain.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Comment extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 300, name = "comment_content")
-    private String commentContent;
-
-    @Column(length = 300, name = "parent_content")
-    private String parentContent;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "event_id")
-    private Event event;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-}
