+++ src/main/java/careerfestival/career/domain/Event.java
@@ -2,9 +2,8 @@ package careerfestival.career.domain;
 
 
 import careerfestival.career.domain.common.BaseEntity;
-import careerfestival.career.domain.mapping.EventImage;
-import careerfestival.career.domain.mapping.EventKeyword;
-import careerfestival.career.domain.mapping.Participate;
+import careerfestival.career.domain.enums.Category;
+import careerfestival.career.domain.mapping.*;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -25,8 +24,6 @@ public class Event extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-
-
     // 모집 시작일 & 모집 종료일
     @Column(nullable = false, name = "recruitment_start")
     private LocalDateTime recruitmentStart;
@@ -46,10 +43,15 @@ public class Event extends BaseEntity {
     private LocalDateTime eventStart;
     @Column(nullable = false, name = "event_end")
     private LocalDateTime eventEnd;
-    @Column(nullable = false, length = 300, name = "link")
+    @Column(length = 300, name = "link")
     private String link;
     @Column(nullable = false, length = 200, name = "event_content")
     private String eventContent;
+
+    // 행사 정보이미지 들어가야함
+
+    @Column(nullable = false, length = 40, name = "event_cost")
+    private String eventCost;
 
     //행사 주소
     @Column(nullable = false, length = 40, name ="address")
@@ -57,22 +59,18 @@ public class Event extends BaseEntity {
     @Column(nullable = false, length = 40, name ="spec_address")
     private String specAddress;
 
-
-
-    //checkpoint 주최자랑 연결?
-    //회원 id 연결 주최자와
     @Column(nullable = false, length = 20, name = "manager_name")
     private String managerName;
-    @Column(nullable = false, length = 20, name = "manager_email")
+    @Column(length = 20, name = "manager_email")
     private String managerEmail;
     @Column(columnDefinition = "INT DEFAULT 0")
     private int hits;
 
+    @Enumerated(EnumType.STRING)
+    private Category category;
+
     @Column(length = 300, name = "event_etc_detail")
     private String eventEtcDetail;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Category category;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Region region;
@@ -82,9 +80,6 @@ public class Event extends BaseEntity {
 
     @OneToMany(mappedBy = "event", cascade = CascadeType.ALL)
     private List<EventImage> eventImg = new ArrayList<>();
-
-    @OneToMany(mappedBy = "event", cascade = CascadeType.ALL)
-    private List<EventKeyword> eventKeyword = new ArrayList<>();
 
     @OneToMany(mappedBy = "event", cascade = CascadeType.ALL)
     private List<Comment> comment = new ArrayList<>();
