+++ src/main/java/careerfestival/career/domain/Record.java
@@ -1,12 +1,12 @@
 package careerfestival.career.domain;
 
 import careerfestival.career.domain.common.BaseEntity;
+import careerfestival.career.domain.enums.Category;
+import careerfestival.career.domain.enums.KeywordName;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDate;
-import careerfestival.career.domain.mapping.RecordKeyword;
-import java.util.ArrayList;
 import java.util.List;
 
 @Entity
@@ -20,13 +20,10 @@ public class Record extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-
     @Column(nullable = false, length = 20, name = "event_name")
     private String eventName;
     @Column(nullable = false, name = "event_date")
     private LocalDate eventDate;
-    @OneToMany(mappedBy = "record", cascade = CascadeType.ALL)
-    private List<RecordKeyword> recordKeywords = new ArrayList<>();
     @Column(nullable = false, length = 300, name = "event_description")
     private String eventDescription;
     @Column(length = 300, name = "networking_name")
@@ -36,9 +33,8 @@ public class Record extends BaseEntity {
     @Column(length = 300, name = "record_etc_detail")
     private String recordEtcDetail;
 
-    // 행사유형
-    @ManyToOne(fetch = FetchType.LAZY)
     private Category category;
+    private List<KeywordName> keywordName;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
