+++ src/main/java/careerfestival/career/domain/enums/Category.java
@@ -1,0 +1,6 @@
+package careerfestival.career.domain.enums;
+
+public enum Category {
+    강연세미나, 학술대회,
+    전시박람회, 기타
+}
