+++ src/main/java/careerfestival/career/domain/mapping/Comment.java
@@ -1,0 +1,54 @@
+package careerfestival.career.domain.mapping;
+
+import careerfestival.career.domain.Event;
+import careerfestival.career.domain.User;
+import careerfestival.career.domain.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Comment extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 300, name = "comment_content")
+    private String commentContent;
+
+    @Column(length = 300, name = "parent_content")
+    private String parentContent;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "event_id")
+    private Event event;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @Entity
+    @Getter
+    @Setter
+    @Builder
+    @NoArgsConstructor(access = AccessLevel.PUBLIC)
+    @AllArgsConstructor(access = AccessLevel.PROTECTED)
+    public static class Organizer extends BaseEntity {
+        @Id
+        @GeneratedValue(strategy = GenerationType.IDENTITY)
+        private Long id;
+
+
+        @Column(nullable = false, length = 20, name = "host_name")
+        private String organizerName;
+
+        @Column(nullable = false, length = 300, name = "profile_img")
+        private String organizerImg;
+
+        @ManyToOne(fetch = FetchType.LAZY)
+        private User user;
+    }
+}
