--- src/main/java/careerfestival/career/domain/mapping/EventKeyword.java
@@ -1,25 +1,0 @@
-package careerfestival.career.domain.mapping;
-
-import careerfestival.career.domain.Event;
-import careerfestival.career.domain.common.BaseEntity;
-import careerfestival.career.domain.enums.KeywordName;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class EventKeyword extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private KeywordName keywordName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Event event;
-}
