+++ src/main/java/careerfestival/career/domain/mapping/Follow.java
@@ -1,5 +1,6 @@
-package careerfestival.career.domain;
+package careerfestival.career.domain.mapping;
 
+import careerfestival.career.domain.User;
 import careerfestival.career.domain.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
@@ -7,21 +8,14 @@ import lombok.*;
 @Entity
 @Getter
 @Setter
-@Builder
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder
 
-public class Host extends BaseEntity {
+public class Follow extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-
-    @Column(nullable = false, length = 20, name = "host_name")
-    private String hostName;
-
-    @Column(nullable = false, length = 300, name = "profile_img")
-    private String profileImg;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
