+++ src/main/java/careerfestival/career/record/controller/RecordController.java
@@ -17,6 +17,7 @@ import java.util.List;
 public class RecordController {
     private final RecordService recordService;
 
+    //
     @PostMapping("/lecture-seminar/{userId}")
     public ResponseEntity<Void> recordLectureSeminar(@PathVariable("userId") Long userId, @RequestBody RecordLectureSeminarDto recordLectureSeminarDto) {
         try {
@@ -27,9 +28,8 @@ public class RecordController {
         }
     }
 
-    // User가 Keyword 테이블에서 선택한다고 했을 때, Keyword_id를 넘겨주는 방법은 어떤지? (어차피 Keyword는 정해져 있으니까?)
-    @PostMapping("/etc")
-    public ResponseEntity<Void> recordEtc(@RequestParam("userId") Long userId, @RequestBody RecordEtcDto recordEtcDto) {
+    @PostMapping("/etc/{userId}")
+    public ResponseEntity<Void> recordEtc(@PathVariable("userId") Long userId, @RequestBody RecordEtcDto recordEtcDto) {
         try {
             recordService.recordEtc(userId, recordEtcDto);
             return new ResponseEntity<>(HttpStatus.OK);
@@ -39,14 +39,13 @@ public class RecordController {
     }
 
     // 메인페이지 1차 구현 완료
-    @GetMapping("")
-    public ResponseEntity<List<RecordMainResponseDto>> getRecordsByUserId(@RequestParam("userId") Long userId) {
+    @GetMapping("/{userId}")
+    public ResponseEntity<List<RecordMainResponseDto>> getRecordsByUserId(@PathVariable("userId") Long userId) {
         try {
             List<RecordMainResponseDto> recordMainResponses = recordService.getRecordsByUserId(userId);
             return ResponseEntity.ok(recordMainResponses);
         } catch (IllegalArgumentException e) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-
     }
 }
