+++ src/main/java/careerfestival/career/record/dto/RecordEtcDto.java
@@ -1,7 +1,6 @@
 package careerfestival.career.record.dto;
 
 import careerfestival.career.domain.Record;
-import careerfestival.career.domain.mapping.RecordKeyword;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,7 +16,6 @@ import java.util.List;
 public class RecordEtcDto {
     private String eventName;
     private LocalDate eventDate;
-    private List<RecordKeyword> recordKeyword;
     private String networkingName;
     private String networkingContact;
 
@@ -26,7 +24,6 @@ public class RecordEtcDto {
         return Record.builder()
                 .eventName(eventName)
                 .eventDate(eventDate)
-                .recordKeywords(recordKeyword)
                 .networkingName(networkingName)
                 .networkingContact(networkingContact)
                 .build();
