+++ src/main/java/careerfestival/career/record/service/RecordService.java
@@ -5,7 +5,6 @@ import careerfestival.career.domain.User;
 import careerfestival.career.record.dto.RecordEtcDto;
 import careerfestival.career.record.dto.RecordLectureSeminarDto;
 import careerfestival.career.record.dto.RecordMainResponseDto;
-import careerfestival.career.repository.RecordKeywordRepository;
 import careerfestival.career.repository.RecordRepository;
 import careerfestival.career.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,22 +17,19 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class RecordService {
     private final RecordRepository recordRepository;
-    private final RecordKeywordRepository recordKeywordRepository;
     private final UserRepository userRepository;
 
     public void recordLectureSeminar(Long userId, RecordLectureSeminarDto recordLectureSeminarDto) {
-
-        // 전달받은 userId로 연관관계 매핑된 User 테이블에 맞춰서 DB에 저장하기
+        // 이미지 첨부 및 글자 수 제한 적용 필요
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("User not found with id: " + userId));
 
         Record record = recordLectureSeminarDto.toEntity();
+        record.setCategory(recordLectureSeminarDto.getCategory());
+        record.setKeywordName(recordLectureSeminarDto.getKeywordName());
         record.setUser(user);
-
         recordRepository.save(record);
     }
-
-
 
     public void recordEtc(Long userId, RecordEtcDto recordEtcDto) {
 
@@ -43,7 +39,6 @@ public class RecordService {
         Record record = recordEtcDto.toEntity();
         record.setUser(user);
         recordRepository.save(record);
-        recordKeywordRepository.saveAll(record.getRecordKeywords());
         // Record 테이블에만 저장된 상태
     }
